pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-eval-pipeline"
version = "0.1.0"
description = "A declarative YAML-based framework for testing and comparing LLM prompts"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pandas>=2.0.0",
    "jinja2>=3.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
prompt-eval = "prompt_eval.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/prompt_eval"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
